refactor(notificationlist): move to recipe
