+++ .storybook/recipes/NotificationListPopover/NotificationListPopover.module.css
@@ -1,6 +1,17 @@
+@import '../../../src/design-tokens/mixins.css';
+
 /*------------------------------------*\
     # NOTIFICATION LIST POPOVER
 \*------------------------------------*/
+
+/**
+* Notification list contains notification items, a count of unread notifications, and a button to mark all notifications as read
+*/
+.notification-list {
+  @mixin reset;
+  display: flex;
+  flex-direction: column;
+}
 
 .notifications-list-popover {
   padding: 0;
@@ -11,4 +22,98 @@
   display: flex;
   justify-content: space-between;
   padding: var(--eds-size-1) var(--eds-size-2);
+}
+
+/**
+* Individual notification list item
+*/
+.notification-list__item {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  padding: var(--eds-size-2);
+  background-color: var(--eds-theme-color-background-brand-primary-default);
+  border: 1px solid var(--eds-theme-color-border-neutral-default);
+}
+
+/**
+* Notification List- is read 
+* 
+* When a list item is marked as read, this class is added to enable "is read" styles.
+*/
+
+.notification-list__is-read {
+  background-color: var(--eds-theme-color-background-neutral-default);
+}
+/* 
+* Notification list item content
+* 
+* The wrapper for the text.
+*/
+.notification-list__content {
+  color: var(--eds-theme-color-text-neutral-default);
+  @mixin eds-theme-typography-body-text-xs;
+}
+
+/*
+* Notification list item title link
+* 
+* Main text title of the notification.
+*/
+.notification-list__link {
+  color: var(--eds-theme-color-text-link-brand);
+  @mixin eds-theme-typography-body-text-sm;
+  text-decoration: none;
+
+  /**
+  *  Notification List Item (is read)
+  * 
+  * When isRead === true, the title link text changes to grey.
+  */
+  .notification-list__is-read & {
+    color: var(--eds-theme-color-text-neutral-default);
+  }
+}
+
+/*
+* Notification list item subtitle
+* 
+* Contains the created date (relative) and the originating source of the notification.
+
+* Notification list item button
+* 
+* Indicator that displays the item's read/not read status in a <button> element;
+* when clicked, it toggles between read/not read statuses.
+*/
+.notification-list__button {
+  position: relative;
+  top: var(--eds-size-1);
+  appearance: none;
+  border: none;
+  outline: none;
+  padding: 0;
+  flex-shrink: 0;
+  border-radius: var(--eds-border-radius-md);
+  height: 0.5rem;
+  width: 0.5rem;
+  background-color: var(--eds-theme-color-border-link-brand);
+  cursor: pointer;
+
+  &:focus-visible {
+    @mixin focus;
+  }
+
+  @supports not selector(:focus-visible) {
+    &:focus {
+      @mixin focus;
+    }
+  }
+}
+
+/*
+  * Notification Button List Item (is read)
+  */
+.notification-list__is-read > .notification-list__button {
+  background-color: var(--eds-theme-color-background-neutral-default);
+  border: 1px solid var(--eds-theme-color-text-neutral-default);
 }
