--- .storybook/recipes/NotificationListPopover/index.ts
@@ -1,1 +1,0 @@
-export { NotificationListPopover as default } from './NotificationListPopover';
