--- src/components/NotificationList/NotificationList.test.ts
@@ -1,6 +1,0 @@
-import { generateSnapshots } from '@chanzuckerberg/story-utils';
-import * as stories from './NotificationList.stories';
-
-describe('<NotificationList />', () => {
-  generateSnapshots(stories);
-});
