--- src/components/NotificationList/index.ts
@@ -1,1 +1,0 @@
-export { NotificationList as default } from './NotificationList';
