--- src/components/NotificationListItem/index.ts
@@ -1,2 +1,0 @@
-export { NotificationListItem as default } from './NotificationListItem';
-export type { NotificationListItemProps } from './NotificationListItem';
