+++ src/index.ts
@@ -117,7 +117,5 @@ export type { BannerProps } from './components/Banner';
 // export { default as Header } from './components/Header';
 // export { default as Main } from './components/Main';
 // export { default as NavContainer } from './components/NavContainer';
-// export { default as NotificationList } from './components/NotificationList';
-// export { default as NotificationListItem } from './components/NotificationListItem';
 // export { default as PrimaryNav } from './components/PrimaryNav';
 // export { default as PrimaryNavItem } from './components/PrimaryNavItem';
